% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGalaxyTool.R
\name{CreateFunctionalTest}
\alias{CreateFunctionalTest}
\title{Create functional test output for tool}
\usage{
CreateFunctionalTest(func, functionalTestDirectory = "inst/functionalTests")
}
\arguments{
\item{func}{Function name to create a tool of, as character value.}

\item{functionalTestDirectory}{functional test directory, recommended to leave as default.}
}
\description{
This function creates the functional test output for a given function.
However, if the tool needs input files then these are expected to be present
in the \code{functionalTestDirectory/func} directory. To see which files exactly
are expected, use the \code{\link{GetGalaxyTestParams}} function.
}
\examples{
# Set your working directory to the root of this package
#   Make sure you have cloned this package to run this example.
\dontrun{
CreateFunctionalTest("CreateDESeqData")
}
}
