% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGalaxyTool.R
\name{RewriteToolScript}
\alias{RewriteToolScript}
\title{Rewrite R script generated by \pkg{RGalaxy}}
\usage{
RewriteToolScript(funcName, galaxyHome = "inst/galaxy", toolDir = "mytool")
}
\arguments{
\item{funcName}{}

\item{galaxyHome}{}

\item{toolDir}{}
}
\description{
Rewrite R script generated by \pkg{RGalaxy}
}
