% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGalaxyTool.R
\name{RewriteToolXML}
\alias{RewriteToolXML}
\title{Rewrite XML generated by \pkg{RGalaxy}}
\usage{
RewriteToolXML(funcName, galaxyHome = "inst/galaxy", toolDir = "mytool")
}
\arguments{
\item{funcName}{}

\item{galaxyHome}{}

\item{toolDir}{}
}
\description{
Rewrite XML generated by \pkg{RGalaxy}
}
